I gang med at skifte place ships til sin egen side
