+++ backend/routes/staticRoutes.js
@@ -41,9 +41,14 @@ router.get('/settings', (req, res) => {
     res.sendFile(join(__dirname, '..', '..', 'frontend', 'templates', 'settings.html'));
 });
 
+router.get('/placeShips', (req, res) => {
+    res.sendFile(join(__dirname, '..', '..', 'frontend', 'templates', 'placeShips.html'));
+});
+
 router.get('/gameLobby', (req, res) => {
     res.sendFile(join(__dirname, '..', '..', 'frontend', 'templates', 'gameLobby.html'));
 });
+
 
 router.get('/game', (req, res) => {
     res.sendFile(join(__dirname, '..', '..', 'frontend', 'templates', 'game.html'));
