+++ frontend/scripts/gameLogic/botGame.js
@@ -1,4 +1,5 @@
-import { initializeFields, randomizeShipPlacement, resetShipPlacement } from './game.js';
+import { initializeFields } from './game.js';
+import { randomizeShipPlacement, resetShipPlacement } from './placeShips.js';
 import { getElementById, querySelectorAll } from '../utility/helperFunctions.js';
 import { createShips } from './ships.js';
