+++ frontend/templates/game.html
@@ -27,35 +27,6 @@
         </div>
     </div>
 
-    <div class="functionWrapper">
-        <div class="shipSelectionWrapper center">
-            <div id="carrierSize5" class="ship" draggable="true"></div>
-            <div id="battleshipSize4" class="ship" draggable="true"></div>
-            <div id="submarineSize3" class="ship" draggable="true"></div>
-            <div id="cruiserSize3" class="ship" draggable="true"></div>
-            <div id="destroyerSize2" class="ship" draggable="true"></div>
-        </div>
-        <div class="buttonWrapper">
-            <div id="gameButtonWrapper">
-                <input id="readyButton" type="button" value="Ready">
-            </div>
-            <div id="gameButtonWrapper">
-                <input id="randomizeButton" type="button" value="Randomize">
-            </div>
-            <div id="gameButtonWrapper">
-                <input id="resetButton" type="button" value="Reset">
-            </div>
-        </div>
-    </div>
-
-    <div id="banner" class="banner">
-        <div class="bannerWrapper">
-            <h2>Awaiting Opponent</h2>
-            <h3 id="gameCode"> Game Code: ... </h3>
-            <input id="cancelButton" type="button" value="Cancel">
-        </div>
-    </div>
-
     <script type="module" src="../scripts/utility/audioManager.js"></script>
     <script type="module" src="../scripts/gameLogic/game.js"></script>
